refactor(decisions): pull backfill, consensus and claim generation out of service.go
Motion only; no function body changes.

internal/service/decisions held exactly four files — service.go at 1,663 lines
plus three test files — so the directory conveyed nothing about what was in it.
AGENTS.md describes this package as "decisions/ (trace pipeline)", and backfill
and consensus are not that.

The seam is a caller-graph fact rather than a length observation:

  - backfill.go (BackfillEmbeddings, BackfillOutcomeEmbeddings, embeddingText,
    backfillSpec, backfillBatch, BackfillClaims, RetryFailedClaimEmbeddings,
    isDuplicateKey) is reached only from background loops in the root package.
    Nothing on an HTTP or MCP request path calls any of it.
  - consensus.go (ConsensusScores, ConsensusScoresBatch, cosineSimFloat32) has
    a single external caller, handlers_decisions.go.
  - claims.go (generateClaims) gets its own file because it has one caller in
    the trace path (postTraceAsync) and two in backfill, so filing it under
    either would misstate ownership.

Neither backfill nor consensus shares state with the trace pipeline, which is
why this splits cleanly where the pipeline itself does not.

The trace pipeline stays whole in service.go, deliberately: the error types,
Service and its setters, New, TraceInput/TraceResult, Trace,
AdjudicateConflictWithTrace, prepareTrace, createDecisionTypeAliasAfterCommit,
recordCompletenessRejectionAudit, postTraceAsync, Check, Search,
hydrateAndReScore, the Query family and ResolveOrCreateAgent. The ordering
between the commit, the after-commit alias write and the rejection audit row is
the durable-audit-ordering property this repo reviews for, and splitting it
across files would make that ordering invisible to a reviewer. 1,199 lines is
the right size for that file.

Verified as pure motion rather than asserted: each block was extracted before
the edit and byte-compared against its destination afterwards — all three
identical (171, 89+105 and 92 lines). The package's three test files needed
zero edits, and `go test -list` returns the same 90 test names. Build, vet and
golangci-lint clean; package green under -race.

Import grouping follows .golangci.yml's goimports local-prefixes setting; a
first pass without -local was caught by the linter rather than by review.
