+++ internal/service/decisions/service.go
@@ -11,7 +11,6 @@ import (
 	"errors"
 	"fmt"
 	"log/slog"
-	"math"
 	"strings"
 	"sync"
 	"time"
@@ -21,8 +20,6 @@ import (
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/metric"
 	"go.opentelemetry.io/otel/trace"
-
-	"golang.org/x/sync/errgroup"
 
 	"github.com/ashita-ai/akashi/internal/conflicts"
 	"github.com/ashita-ai/akashi/internal/ctxutil"
@@ -1114,178 +1111,6 @@ func (s *Service) Recent(ctx context.Context, orgID uuid.UUID, filters model.Que
 	})
 }
 
-// ConsensusScores returns agreement and conflict counts for a single decision.
-//
-// Agreement is computed by:
-//  1. Asking Qdrant for the top-50 embedding neighbors of this decision.
-//  2. Fetching their outcome_embeddings from Postgres.
-//  3. Counting neighbors where outcome cosine similarity ≥ 0.75.
-//
-// Conflict count is an index-backed join on scored_conflicts (no ANN needed).
-// Returns 0 for agreement if Qdrant is not available or the decision has no embedding.
-func (s *Service) ConsensusScores(ctx context.Context, decisionID, orgID uuid.UUID) (agreementCount, conflictCount int, err error) {
-	conflictCount, err = s.db.GetConflictCount(ctx, decisionID, orgID)
-	if err != nil {
-		return 0, 0, err
-	}
-
-	cf, ok := s.searcher.(search.CandidateFinder)
-	if !ok {
-		return 0, conflictCount, nil
-	}
-	if s.searcher.Healthy(ctx) != nil {
-		return 0, conflictCount, nil
-	}
-
-	// Fetch the source decision's embeddings for the Qdrant query and pairwise comparison.
-	embMap, err := s.db.GetDecisionEmbeddings(ctx, []uuid.UUID{decisionID}, orgID)
-	if err != nil || len(embMap) == 0 {
-		return 0, conflictCount, nil //nolint:nilerr // missing embedding is not an error
-	}
-	embs := embMap[decisionID]
-	sourceOutcome := embs[1]
-
-	results, err := cf.FindSimilar(ctx, orgID, embs[0].Slice(), decisionID, nil, 50)
-	if err != nil {
-		s.logger.Warn("consensus: qdrant find similar failed", "decision_id", decisionID, "error", err)
-		return 0, conflictCount, nil
-	}
-	if len(results) == 0 {
-		return 0, conflictCount, nil
-	}
-
-	neighborIDs := make([]uuid.UUID, len(results))
-	for i, r := range results {
-		neighborIDs[i] = r.DecisionID
-	}
-	neighborEmbs, err := s.db.GetDecisionEmbeddings(ctx, neighborIDs, orgID)
-	if err != nil {
-		s.logger.Warn("consensus: fetch neighbor embeddings failed", "error", err)
-		return 0, conflictCount, nil
-	}
-
-	for _, id := range neighborIDs {
-		if n, ok := neighborEmbs[id]; ok {
-			if cosineSimFloat32(sourceOutcome.Slice(), n[1].Slice()) >= 0.75 {
-				agreementCount++
-			}
-		}
-	}
-	return agreementCount, conflictCount, nil
-}
-
-// ConsensusScoresBatch returns agreement and conflict counts for multiple decisions.
-// Conflict counts are computed via a batch join on scored_conflicts (no ANN).
-// Agreement counts use Qdrant for ANN; decisions without Qdrant have agreement=0.
-// The returned map includes all input IDs; absent entries have [0,0] scores.
-func (s *Service) ConsensusScoresBatch(ctx context.Context, ids []uuid.UUID, orgID uuid.UUID) (map[uuid.UUID][2]int, error) {
-	result := make(map[uuid.UUID][2]int, len(ids))
-
-	// Conflict counts: index-backed join, no ANN needed.
-	conflictCounts, err := s.db.GetConflictCountsBatch(ctx, ids, orgID)
-	if err != nil {
-		return nil, err
-	}
-	for id, count := range conflictCounts {
-		v := result[id]
-		v[1] = count
-		result[id] = v
-	}
-
-	cf, ok := s.searcher.(search.CandidateFinder)
-	if !ok || s.searcher.Healthy(ctx) != nil {
-		return result, nil
-	}
-
-	// Fetch embeddings for all input decisions (only those with both populated).
-	embMap, err := s.db.GetDecisionEmbeddings(ctx, ids, orgID)
-	if err != nil || len(embMap) == 0 {
-		return result, nil //nolint:nilerr
-	}
-
-	// Per-decision Qdrant queries, parallelized with bounded concurrency.
-	// Collect all neighbor IDs for a single batch Postgres fetch, avoiding
-	// N+1 embedding lookups.
-	neighborResultsByID := make(map[uuid.UUID][]search.Result, len(embMap))
-	allNeighborIDs := make(map[uuid.UUID]struct{})
-	var mu sync.Mutex
-
-	g, gCtx := errgroup.WithContext(ctx)
-	g.SetLimit(10) // bound concurrent Qdrant RPCs
-
-	for id, embs := range embMap {
-		g.Go(func() error {
-			results, qErr := cf.FindSimilar(gCtx, orgID, embs[0].Slice(), id, nil, 50)
-			if qErr != nil {
-				s.logger.Warn("consensus batch: qdrant find similar failed", "decision_id", id, "error", qErr)
-				return nil // non-fatal: skip this decision
-			}
-			mu.Lock()
-			neighborResultsByID[id] = results
-			for _, r := range results {
-				allNeighborIDs[r.DecisionID] = struct{}{}
-			}
-			mu.Unlock()
-			return nil
-		})
-	}
-	if err := g.Wait(); err != nil {
-		return result, nil //nolint:nilerr // Qdrant errors are non-fatal for consensus
-	}
-
-	if len(allNeighborIDs) == 0 {
-		return result, nil
-	}
-
-	// Batch-fetch outcome embeddings for all unique neighbors.
-	neighborIDList := make([]uuid.UUID, 0, len(allNeighborIDs))
-	for id := range allNeighborIDs {
-		neighborIDList = append(neighborIDList, id)
-	}
-	neighborEmbs, err := s.db.GetDecisionEmbeddings(ctx, neighborIDList, orgID)
-	if err != nil {
-		s.logger.Warn("consensus batch: fetch neighbor embeddings failed", "error", err)
-		return result, nil
-	}
-
-	// Compute agreement counts using outcome cosine similarity in Go.
-	for id, embs := range embMap {
-		neighbors := neighborResultsByID[id]
-		agreement := 0
-		for _, r := range neighbors {
-			if n, ok := neighborEmbs[r.DecisionID]; ok {
-				if cosineSimFloat32(embs[1].Slice(), n[1].Slice()) >= 0.75 {
-					agreement++
-				}
-			}
-		}
-		v := result[id]
-		v[0] = agreement
-		result[id] = v
-	}
-
-	return result, nil
-}
-
-// cosineSimFloat32 computes cosine similarity between two float32 vectors.
-// Returns 0 for zero-length or mismatched vectors.
-func cosineSimFloat32(a, b []float32) float64 {
-	if len(a) != len(b) || len(a) == 0 {
-		return 0
-	}
-	var dot, normA, normB float64
-	for i := range a {
-		da, db := float64(a[i]), float64(b[i])
-		dot += da * db
-		normA += da * da
-		normB += db * db
-	}
-	if normA == 0 || normB == 0 {
-		return 0
-	}
-	return dot / (math.Sqrt(normA) * math.Sqrt(normB))
-}
-
 // SemanticSearchAvailable returns true when both Qdrant and a real embedding
 // provider are configured. Used by the config endpoint so the UI can accurately
 // show whether semantic (vector) search is available.
@@ -1371,293 +1196,4 @@ func (s *Service) ResolveOrCreateAgent(ctx context.Context, orgID uuid.UUID, age
 
 	s.logger.Info("auto-registered agent on first trace", "agent_id", agentID, "org_id", orgID)
 	return agent, nil
-}
-
-// BackfillEmbeddings generates embeddings for decisions that were stored without
-// one (e.g. because the embedding provider was noop at trace time). Each decision
-// is embedded, the vector is written to Postgres, and a search outbox entry is
-// queued so the outbox worker can sync it to Qdrant.
-//
-// Returns the number of decisions backfilled. Skips silently if the embedding
-// provider is noop (returns 0, nil).
-func (s *Service) BackfillEmbeddings(ctx context.Context, batchSize int) (int, error) {
-	return s.backfillBatch(ctx, batchSize, backfillSpec{
-		find:  s.db.FindUnembeddedDecisions,
-		text:  embeddingText,
-		write: s.db.BackfillEmbedding,
-		label: "backfill: embedded decisions",
-	})
-}
-
-// BackfillOutcomeEmbeddings populates outcome_embedding for decisions that have
-// embedding but no outcome_embedding (Option B). Returns the number backfilled.
-func (s *Service) BackfillOutcomeEmbeddings(ctx context.Context, batchSize int) (int, error) {
-	return s.backfillBatch(ctx, batchSize, backfillSpec{
-		find:  s.db.FindDecisionsMissingOutcomeEmbedding,
-		text:  func(d storage.UnembeddedDecision) string { return d.Outcome },
-		write: s.db.BackfillOutcomeEmbedding,
-		label: "backfill: outcome embeddings",
-	})
-}
-
-// embeddingText builds the canonical embedding input for a decision (same
-// format used by prepareTrace).
-func embeddingText(d storage.UnembeddedDecision) string {
-	s := d.DecisionType + ": " + d.Outcome
-	if d.Reasoning != nil {
-		s += " " + *d.Reasoning
-	}
-	return s
-}
-
-// backfillSpec parameterizes the shared backfill loop.
-type backfillSpec struct {
-	find  func(ctx context.Context, limit int) ([]storage.UnembeddedDecision, error)
-	text  func(d storage.UnembeddedDecision) string
-	write func(ctx context.Context, id uuid.UUID, orgID uuid.UUID, vec pgvector.Vector) error
-	label string
-}
-
-// backfillBatch probes the embedding provider, finds records needing backfill,
-// embeds them in a single batch, and writes each vector back. Shared by
-// BackfillEmbeddings and BackfillOutcomeEmbeddings.
-func (s *Service) backfillBatch(ctx context.Context, batchSize int, spec backfillSpec) (int, error) {
-	if _, err := s.embedder.Embed(ctx, "probe"); errors.Is(err, embedding.ErrNoProvider) {
-		return 0, nil
-	}
-
-	decs, err := spec.find(ctx, batchSize)
-	if err != nil {
-		return 0, fmt.Errorf("%s: find: %w", spec.label, err)
-	}
-	if len(decs) == 0 {
-		return 0, nil
-	}
-
-	texts := make([]string, len(decs))
-	for i, d := range decs {
-		texts[i] = spec.text(d)
-	}
-
-	vecs, err := s.embedder.EmbedBatch(ctx, texts)
-	if err != nil {
-		return 0, fmt.Errorf("%s: embed batch: %w", spec.label, err)
-	}
-
-	var backfilled int
-	for i, d := range decs {
-		if err := s.validateEmbeddingDims(vecs[i]); err != nil {
-			s.logger.Warn(spec.label+": dimension mismatch, skipping", "decision_id", d.ID, "error", err)
-			continue
-		}
-		if err := spec.write(ctx, d.ID, d.OrgID, vecs[i]); err != nil {
-			s.logger.Warn(spec.label+": update failed", "decision_id", d.ID, "error", err)
-			continue
-		}
-		backfilled++
-	}
-
-	if backfilled > 0 {
-		s.logger.Info(spec.label, "count", backfilled, "batch", len(decs))
-	}
-	return backfilled, nil
-}
-
-// generateClaims extracts claims from an outcome, embeds each, and stores them
-// in the decision_claims table. When an LLM extractor is configured, claims are
-// extracted with categories (finding, recommendation, assessment, status).
-// Otherwise falls back to regex-based SplitClaims (uncategorized).
-// Skips if claims already exist.
-func (s *Service) generateClaims(ctx context.Context, decisionID, orgID uuid.UUID, outcome string) error {
-	// Skip if claims already exist for this decision.
-	exists, err := s.db.HasClaimsForDecision(ctx, decisionID, orgID)
-	if err != nil {
-		return fmt.Errorf("claims: check existing: %w", err)
-	}
-	if exists {
-		return nil
-	}
-
-	type textAndCategory struct {
-		text     string
-		category *string // nil for regex-extracted claims
-	}
-
-	var extracted []textAndCategory
-
-	if s.claimExtractor != nil {
-		llmClaims, err := s.claimExtractor.ExtractClaims(ctx, outcome)
-		if err != nil {
-			// Fall back to regex on LLM failure.
-			s.logger.Warn("claims: LLM extraction failed, falling back to regex",
-				"decision_id", decisionID, "error", err)
-		} else {
-			for _, c := range llmClaims {
-				cat := c.Category
-				extracted = append(extracted, textAndCategory{text: c.Text, category: &cat})
-			}
-		}
-	}
-
-	// Fallback: regex-based extraction (no categories).
-	if len(extracted) == 0 {
-		for _, text := range conflicts.SplitClaims(outcome) {
-			extracted = append(extracted, textAndCategory{text: text, category: nil})
-		}
-	}
-
-	if len(extracted) == 0 {
-		return nil
-	}
-
-	// Embed all claims in a single batch call.
-	texts := make([]string, len(extracted))
-	for i, e := range extracted {
-		texts[i] = e.text
-	}
-	vecs, err := s.embedder.EmbedBatch(ctx, texts)
-	if err != nil {
-		return fmt.Errorf("claims: embed batch: %w", err)
-	}
-
-	// Build claim records.
-	claims := make([]storage.Claim, 0, len(extracted))
-	for i, e := range extracted {
-		if i >= len(vecs) {
-			break
-		}
-		if err := s.validateEmbeddingDims(vecs[i]); err != nil {
-			s.logger.Warn("claims: dimension mismatch, skipping claim", "decision_id", decisionID, "claim_idx", i, "error", err)
-			continue
-		}
-		emb := vecs[i]
-		claims = append(claims, storage.Claim{
-			DecisionID: decisionID,
-			OrgID:      orgID,
-			ClaimIdx:   i,
-			ClaimText:  e.text,
-			Category:   e.category,
-			Embedding:  &emb,
-		})
-	}
-
-	if len(claims) == 0 {
-		return nil
-	}
-
-	if err := s.db.InsertClaims(ctx, claims); err != nil {
-		return fmt.Errorf("claims: insert: %w", err)
-	}
-	method := "regex"
-	if s.claimExtractor != nil && extracted[0].category != nil {
-		method = "llm"
-	}
-	s.logger.Debug("claims: generated", "decision_id", decisionID, "count", len(claims), "method", method)
-	return nil
-}
-
-// BackfillClaims generates sentence-level claim embeddings for decisions that
-// have embeddings but no claims yet. Returns the number of decisions processed.
-func (s *Service) BackfillClaims(ctx context.Context, batchSize int) (int, error) {
-	if _, err := s.embedder.Embed(ctx, "probe"); errors.Is(err, embedding.ErrNoProvider) {
-		return 0, nil
-	}
-
-	refs, err := s.db.FindDecisionIDsMissingClaims(ctx, batchSize)
-	if err != nil {
-		return 0, fmt.Errorf("backfill claims: find: %w", err)
-	}
-	if len(refs) == 0 {
-		return 0, nil
-	}
-
-	var backfilled int
-	for _, ref := range refs {
-		select {
-		case <-ctx.Done():
-			return backfilled, ctx.Err()
-		default:
-		}
-		// Fetch the decision outcome.
-		d, err := s.db.GetDecisionForScoring(ctx, ref.ID, ref.OrgID)
-		if err != nil {
-			s.logger.Warn("backfill claims: get decision failed", "decision_id", ref.ID, "error", err)
-			continue
-		}
-		if err := s.generateClaims(ctx, ref.ID, ref.OrgID, d.Outcome); err != nil {
-			s.logger.Warn("backfill claims: generate failed", "decision_id", ref.ID, "error", err)
-			continue
-		}
-		backfilled++
-	}
-
-	if backfilled > 0 {
-		s.logger.Info("backfill: claims generated", "count", backfilled, "batch", len(refs))
-	}
-	return backfilled, nil
-}
-
-// RetryFailedClaimEmbeddings re-attempts claim embedding generation for decisions
-// that failed previously and are eligible for retry (exponential backoff, capped
-// at maxAttempts). On success, clears the failure state and triggers conflict
-// scoring. On failure, increments the attempt counter for longer backoff.
-// Returns the number of decisions successfully retried.
-func (s *Service) RetryFailedClaimEmbeddings(ctx context.Context, batchSize, maxAttempts int) (int, error) {
-	if _, err := s.embedder.Embed(ctx, "probe"); errors.Is(err, embedding.ErrNoProvider) {
-		return 0, nil
-	}
-
-	refs, err := s.db.FindRetriableClaimFailures(ctx, maxAttempts, batchSize)
-	if err != nil {
-		return 0, fmt.Errorf("retry claims: find: %w", err)
-	}
-	if len(refs) == 0 {
-		return 0, nil
-	}
-
-	var retried int
-	for _, ref := range refs {
-		select {
-		case <-ctx.Done():
-			return retried, ctx.Err()
-		default:
-		}
-
-		d, err := s.db.GetDecisionForScoring(ctx, ref.ID, ref.OrgID)
-		if err != nil {
-			s.logger.Warn("retry claims: get decision failed", "decision_id", ref.ID, "error", err)
-			continue
-		}
-
-		if err := s.generateClaims(ctx, ref.ID, ref.OrgID, d.Outcome); err != nil {
-			s.logger.Warn("retry claims: generate failed", "decision_id", ref.ID, "error", err)
-			s.claimEmbeddingFailures.Add(ctx, 1, metric.WithAttributes(
-				attribute.Int("attempt_number", ref.Attempts+1),
-			))
-			if markErr := s.db.MarkClaimEmbeddingFailed(ctx, ref.ID, ref.OrgID); markErr != nil {
-				s.logger.Error("retry claims: failed to mark failure", "decision_id", ref.ID, "error", markErr)
-			}
-			continue
-		}
-
-		if err := s.db.ClearClaimEmbeddingFailure(ctx, ref.ID, ref.OrgID); err != nil {
-			s.logger.Error("retry claims: failed to clear failure state", "decision_id", ref.ID, "error", err)
-		}
-
-		if s.conflictScorer != nil {
-			s.conflictScorer.ScoreForDecision(ctx, ref.ID, ref.OrgID)
-		}
-		retried++
-	}
-
-	if retried > 0 {
-		s.logger.Info("retry claims: succeeded", "count", retried, "batch", len(refs))
-	}
-	return retried, nil
-}
-
-// isDuplicateKey delegates to the storage backend to check for unique constraint
-// violations (Postgres 23505, SQLite CONSTRAINT_UNIQUE, etc.).
-func (s *Service) isDuplicateKey(err error) bool {
-	return s.db.IsDuplicateKey(err)
 }
